feat(distinct): add higher-order lettable version of distinct
